[backend] update evaluation result structure
